requirements.txt
altgraph==0.16.1
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
future==0.17.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
macholib==1.11
MarkupSafe==1.1.1
openapi-codec==1.3.2
pefile==2019.4.18
PyInstaller==3.4
pytz==2019.1
pywin32-ctypes==0.2.0
requests==2.22.0
simplejson==3.16.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.2
